requirements.txt
numpy>=2.0.2
pillow>=9.2.0
scikit-learn>=1.6.0
pygame-ce>=2.5.2